moved the login, and token decoders and encoders implementation in their own source files
